Refactored AbstractSyntaxTree file
Merged the AST class into the language.rb file as there were some circular references between the two files. I haven't ironed out the best way to separate these large ruby source files into smaller files without causing import issues.
